Refactor code to better separate functionality
Now the call API functions can be used to set the values of weather
and gif data. They can also be used to create promises on whose
resolution we can access the values in weather and gif data. This
can allow exporting of data from this module allowing this module
to only send data.